Jl/better validation

+++ src/controllers/survey123.js
@@ -101,6 +101,10 @@ export const uploadCsv = async (filename) => {
   // spread out the operation into sequential deletes and inserts
   const bulkOp = docs.flatMap(deleteInsert).filter((obj) => !!obj);
 
+  if (!bulkOp.length) {
+    throw newError(RESPONSE_TYPES.BAD_REQUEST, 'no valid data');
+  }
+
   const insertOp = bulkOp.filter(({ insertOne }) => !!insertOne);
   const deleteOp = bulkOp.filter(({ deleteMany }) => !!deleteMany);
 
@@ -110,7 +114,7 @@ export const uploadCsv = async (filename) => {
   console.log(`successfully parsed ${rowCount} rows from csv upload`);
 
   // run entire pipeline
-  // don't throw the error here since we want the webhook to return 200 immediately
+  // don't throw the error here since we want to return 200 immediately
   // also don't await it for the same purpose; run pipeline in background
   runPipelineAll().catch(console.error);
 
@@ -181,7 +185,7 @@ export const uploadSurvey123FromWebhook = async (rawData) => {
   const deleteInsertRes = await UnsummarizedTrappingModel.bulkWrite(deleteInsertOp, { ordered: true });
 
   // run entire pipeline
-  // don't throw the error here since we want the webhook to return 200 immediately
+  // don't throw the error here since we want to return 200 immediately
   // also don't await it for the same purpose; run pipeline in background
   runPipelineAll().catch(console.error);
